Unify all results notifications to use new component CippCallout which combines CCallout and CAlert but is dismissible.  Fix issues with notification not appearing when running a second time. Fix issue with password config not working.
